Add component of custom controls to simulator and video player, fix simulator issue on IOS
