+++ src/libs/plyr/plyr-options.ts
@@ -1,0 +1,11 @@
+import type Plyr from 'plyr';
+
+export const plyrOptions: Plyr.Options = {
+    storage: {
+        enabled: false,
+    },
+
+    fullscreen: {
+        iosNative: true,
+    },
+};
