+++ src/utils/navigator.utils.ts
@@ -1,0 +1,6 @@
+export const isIOS = () => {
+    const userAgent = navigator.userAgent;
+    const isIOS = /iPhone|iPad|iPod/i.test(userAgent);
+    const isSafari = /^((?!chrome|android).)*safari/i.test(userAgent);
+    return isIOS && isSafari;
+};
